refactor\!: make all implementation modules private (_<name>)
Rename every public module to a leading-underscore name (agenda -> _agenda,
common -> _common, courses -> _courses, ... session -> _session, etc.) and their
.pyi stubs to match, so the only supported import path is the package root
(`from smartschool import ...`). All public names are re-exported from __init__;
internal modules import each other directly. Updates all internal imports and the
test mock-patch target strings.

BREAKING CHANGE: importing from smartschool.<submodule> no longer works; import
from smartschool directly. Major version bump.
