+++ src/smartschool/_exceptions.py
